fix: resolve button removal issue
Buttons are displayed as part of the same bubble - within the same,
bubble, we loop over the array of buttons presented. However, because
we use the same variable name 'index' to loop over the array of buttons
as we do over the array of bubbles, we are unable to delete the correct
bubble from the array. This commit fixes this issue.