fix: correct errors in parseMarkdown
Since we use a markdown converter to convert to HTML, we need to
account for the difference in treatment between list items and text
blocks - text blocks are wrapped in <p> tags whereas list items are
wrapped in either <ol> or <ul>

This means that for the most part, we can rely on the difference in tags
between <p> and <ol> or <ul> to generate the line breaks. However, if
two consecutive text blocks appear and are separate by line breaks in
markdown, then they will appear as two consecutive <p> blocks and there
will not be appropriate line spacing between them.

To resolve this problem, we detect specifically for the case where there
are two text blocks in a row, in which case we swap our the line breaks
for break tags <br>